backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

AUTH0_DOMAIN = os.getenv("AUTH0_DOMAIN")
AUTH0_ISSUER = f"https://{AUTH0_DOMAIN}/"
AUTH0_JWKS_URL = f"{AUTH0_ISSUER}.well-known/jwks.json"
AUTH0_AUDIENCE = os.getenv("AUTH0_AUDIENCE")

ENCRYPTION_KEY = os.getenv("ENCRYPTION_KEY")
if not ENCRYPTION_KEY:
    raise ValueError("ENCRYPTION_KEY missing in .env")

DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./vault.db")
