requirements.txt
fastapi 
uvicorn 
sqlalchemy 
python-dotenv 
cryptography 
pyjwt 
requests 
python-multipart 
jose[cryptography]
